+++ Encoding/VCEEnc.vb
@@ -239,6 +239,10 @@ Public Class VCEEnc
             Select Case Decoder.ValueText
                 Case "avs"
                     sourcePath = p.Script.Path
+
+                    If includePaths AndAlso FrameServerHelp.IsAviSynthPortableUsed Then
+                        ret += " --avsdll " + Package.AviSynth.Path.Escape
+                    End If
                 Case "qs"
                     sourcePath = "-"